requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
idna==3.3
    # via requests
pydantic==1.9.1
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
twitter==1.19.3
    # via -r requirements.in
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.10
    # via requests
